refactor(fleet): split the tests module out of fleet.rs, and retighten the size baseline
fleet.rs crossed its grandfathered ceiling with the task-timeout work; the
house fix is the pipeline.rs shape — tests in a child module file reaching
private internals via super::* — which takes fleet.rs to ~1000 lines and
retires its baseline exemption entirely. The remaining bumps are surgical
growth in already-grandfathered files (pipeline.rs +37, command_deck.rs +38,
two test files +18 each) and zai.rs +4, which was already over its ceiling
on the base commit. Also de-links private consts from public rustdoc
(rustdoc -D warnings).
