+++ scripts/file-size-baseline.txt
@@ -15,20 +15,20 @@
 1887 bench/terminal_bench_analysis/tb21_evidence_contract.py
 4173 bench/terminal_bench_analysis/tests/test_tb21_analysis.py
 1615 bench/terminal_bench_analysis/tests/test_tb21_evidence_contract.py
-2350 stella-cli/src/agent.rs
-1794 stella-cli/src/agent_tests.rs
+2349 stella-cli/src/agent.rs
+1812 stella-cli/src/agent_tests.rs
 1576 stella-cli/src/candidate_ws.rs
-4445 stella-cli/src/command_deck.rs
+4483 stella-cli/src/command_deck.rs
+1526 stella-cli/src/memory/tests.rs
 2129 stella-core/src/bus.rs
 2563 stella-core/src/driver.rs
 3543 stella-core/src/driver/tests.rs
-1849 stella-fleet/src/fleet.rs
 1628 stella-model/src/anthropic/tests.rs
 2060 stella-model/src/openai.rs
-1520 stella-model/src/zai.rs
+1524 stella-model/src/zai.rs
 1773 stella-model/src/zai/tests.rs
-3381 stella-pipeline/src/pipeline.rs
-2440 stella-pipeline/src/pipeline/tests.rs
+3418 stella-pipeline/src/pipeline.rs
+2458 stella-pipeline/src/pipeline/tests.rs
 2784 stella-protocol/src/event.rs
 2078 stella-store/src/lib.rs
 2261 stella-store/src/tests.rs
